bootloader/lk{,env}: use xerrors to make Load() err observable as ErrNotExist
Previously we would never continue in SetBootVars if the file doesn't exist
because we used fmt.Errorf which does not preserve the error type such that
os.IsNotExist would return true.

Now, we can continue to provide a customized error message from Load(), as well
as make SetBootVars continue and just write out a new file even if we can't read
the lkenv file in Load().

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>