snap-confine: make host /etc/ssl available for snaps on classic
This commit will make/etc/ssl avaialble for snaps if the host
is Ubuntu/Debian and if the base is "core*". This is the first
and simple step for supporting ssl certificates for snaps.

To support non-Ubuntu bases or non-Ubuntu/Debian classic systems
or Ubuntu Core more work is required.