many: rename disks.FindMatching... to FindMatching...WithFsLabel and err types
This is to accommodate a future FindMatching...WithPartLabel. Additionally, an
internal refactor to support storing additional properties for partitions when
we search a disk is implemented to allow for the additional future method.

The error type now is also more versatile to allow specifying what part failed,
searching for a partition by filesystem label or by partition label.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>