bootloader/lk.go: fix handling of os.ErrNotExist everywhere, add comments
We should be using xerrors.Is() everywhere here, since the error returned is
wrapped.

However, for specifically RemoveKernelAssets, it's not a bug that we would skip
over os.ErrNotExist, but it would result in a more confusing error message - we
should try to give better error messages, and returning early with
os.ErrNotExist is more useful to the user than the existing error around not
being able to find the specified kernel blob.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>