many: rename disks.FindMatching... to FindMatching...WithFsLabel and err type (#9720)
* many: rename disks.FindMatching... to FindMatching...WithFsLabel and err types

This is to accommodate a future FindMatching...WithPartLabel. Additionally, an
internal refactor to support storing additional properties for partitions when
we search a disk is implemented to allow for the additional future method.

The error type now is also more versatile to allow specifying what part failed,
searching for a partition by filesystem label or by partition label.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* many: rename FindMatching...FromFsLabel to FindMatching...WithFsLabel

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* osutil/disks/disks.go: demote TODO:UC20 to normal TODO's

These things aren't blocking UC20 development and are just generic things we
would like to do $SOMEDAY for better error reporting.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* osutil/disks/disks_linux.go: adjust comment about optional partition labels

Thanks to @bboozzoo for clarifying this for me. Partition labels are optional on
GPT partitions.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* osutil/disks/disks.go: fix typo

Thanks to @bboozzoo for catching this.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* osutil/disks/disks_linux.go: add some TODOs about things to cleanup/simplify

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* osutil/disks/disks_linux.go: add comments about the encoding of the udev props

Thanks to @bboozzoo for the suggestion.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>

* osutil/disks/mockdisk.go: fix typo

Thanks to @bboozzoo again for spotting all my typos :-)

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>