bootloader/lk: use specific backup file for lkenv
For UC20 runtime, the backup file will be a different partition, and since we
are using partition uuid symlinks instead of partition labels, we can't just
append "bak" to the primary env file, and instead need to manually provide the
backup file.

Also add a test which validates that the backup env file for run mode on uc20 is
handled appropriately and updated.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>