bootloader/lk: update handling of os.ErrNotExist for various cases
The only case where we can really proceed without having an initial env file is
at prepare image time when the bootloader env file doesn't exist, as we could be
writing an empty/new one. All other cases we need the file to exist as a
pre-requisite, if only for accurate error reporting.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>