interfaces/apparmor: do not fail during initialization when there is no AppArmor profile for snap-confine
Some distros support AppArmor in the kernel and have the necessary userspace
tools, but the snapd package (specifically the snap-confine binary) is built
without AppArmor support. In this case, the AppArmor profiles files will not be
present on the host.

Should such distro run with /home on NFS, or overlay set up for /, the apparmor
backend initialization would fail on trying to update the profile files, this
preventing snapd from operating at all. In particular, this happens on openSUSE
Leap 15.x where the snapd package is built with AppArmor support disabled.

See https://forum.snapcraft.io/t/snapd-doesnt-start-on-opensuse-leap-15-1/13710
for more details.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>