secboot/secboot_tpm.go: increase frequency we check for hook completion
The resolution of how often we check directly affects boot speed, since we need
to wait at least one cycle if the file is not created instantly, so rather than
always hold up boot for at least one second, make the time period between checks
smaller but not too small such that we are hogging CPU cycles checking for a
file. 20 Hz seems a reasonable compromise for this.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>