secboot/secboot_tpm.go: manually delete and recreate the std{in,out,err} streams
We want to be robust and secure, so to guard against possible bugs where the
files weren't properly deleted and/or have the wrong permissions, we always
should delete and re-create the stream files with the right permissions before
executing the hook itself.

Also add a test with a hook implementation which verifies the files have the 
right permissions and content that we expect.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>