o/snapstate: add Flags.OverrideToDevModeViaModel , use in ensureInstall...
This will be set by devicestate via options in the model assertion, currently
only when the grade is dangerous, but eventually it may be set for non-dangerous
grades via other things in the model assertion.

When this flag is set, it allows a snap to be installed with devmode
confinement, and will force that snap to be installed with confinement devmode.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>