overlord/devicestate/firstboot.go: in dangerous model, allow devmode snaps
This allows a dangerous grade model assertion to have devmode snaps installed,
and furthermore to force them to be installed with devmode confinement. We don't
set flags.DevMode to true here because we only set that to true in conjunction
with OverrideToDevModeViaModel if and only if the snap.yaml has confinement
devmode, i.e. we don't want to force all snaps in a dangerous image to be
devmode, just snaps that are already devmode. We don't have the snap.yaml info
at this point in time, so we can't make that decision here, and so instead that
is done in snapstate via ensureInstallPreconditions from snapstate.InstallPath.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>