Allow to set a specific PROJECT_PATH
In case the PROJECT_PATH is not in $GOHOME/src/github.com/snapcore/snapd
the pkgdb script fails to find the os.query binary and other scripts.

The idea is to allow to define a specific PROJECT_PATH on spread cron
jobs where we clone the snpad code and run tests.