systemd/systemd.go: refactor again the journald data parsing
Actually, if we encounter a list of numbers, it is not a list of runes, it is a
list of bytes, so we have to decode into an array of bytes and then turn that
into a string, possibly with Go escaping invalid UTF-8 with it's internal hex
encoding.

Also, if the error for getting the message from the Log entry is more than just
that the key "MESSAGE" was missing entirely, include the error in the returned
string for clarity's sake.

Additionally, add more test cases here.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>