interfaces/builtin: fix unit test expecting snap-device-helper at /usr/lib/snapd
The unit tests may be run on other systems as part of the package build or while
developing snapd on those systems. On systems where /usr/libexec/snapd is used,
the unit tests fail as they hardcode /usr/lib/snapd/snap-device-helper. We do
not run unit tests on such platform as part of our CI runs, so this went largely
unnoticed.

Instead of mocking the system release info, update the tests to expect correct
path, such that we know that the snapd binary generates proper rules.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>