usersession/autostart: change ~/snap perms to 0700 on startup
Use the userd autostart machinery to change the permissions on ~/snap to protect
all files and data that snaps store in $SNAP_USER_DATA or $SNAP_USER_COMMON.
Using the user session agent means that we are already running as the right user
to perform this fix, and also means that the fix is applied whenever an affected
user logs in if they have desktop snaps that autostart (such as snap-store).

This is part of the fix, the other part of the fix will be making `snap run`
perform the same restriction in case a user does not use autostart desktop
snaps, their user snap home directories will get fixed when they first use a
snap.

Partially fixes: https://bugs.launchpad.net/snapd/+bug/1910298

Partially fixes: CVE-2021-3155

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>