cmd/snap-repair/runner.go: add SNAP_SYSTEM_MODE to env of repair runner
This can be used by repair assertions to detect what mode they are running in,
if they are meant to be used in multiple modes. Eventually we may need to
include more information here, but note that repair assertion scripts run
unconfined, so they could also parse degraded.json themselves in the meantime
for example.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>