vendor: fix versions of github.com/snapcore/secboot and github.com/canonical/go-tpm2
The latest master branches of both packages are incompatible due to API changes
in go-tpm2 that aren't supported in the secboot package. At the same time,
secboot introduced an API change that breaks snapd.

Downgrade go-tpm2 version to revision listed in secboot's vendor file which
points to this commit:
https://github.com/canonical/go-tpm2/commit/fb781d04d0dea72f916cc68b637b7be6bcb3ccd4

Downgrade secboot to this commit:
https://github.com/snapcore/secboot/commit/027522efe9c7dc36adde13ea8ab030d3dfa34267
which includes https://github.com/snapcore/secboot/pull/108 that made the TPM
the boot policy less strict.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>