tests/regression/lp-1910456: also test that classic snaps are fixed too
We have classic snaps such as eks and microk8s which are container management
snaps and are also vulnerable to CVE-2020-27352, and as such should be fixed as
well. The fix we have in place works equally well for classic snaps as strict
snaps, so just expand the test matrix here to also test that an equivalent
classic snap is fixed too.

This part of the test though should only be run on classic distros, not on 
Ubuntu Core, so some logic around that has to be specified.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>