interfaces/apparmor: Support using a snapd internal apparmor_parser
Instead of exec'ing apparmor_parser from PATH, use the FindAppArmorParser()
method provided by sandbox/apparmor to locate the apparmor_parser binary so
that we use the same one as is indexed by system-key via the parser
features etc. In this case, when using this internal apparmor_parser we
ensure it is executed with additional command-line arguments to also use
the snapd internal apparmor parser configuration file and profile
abstractions etc so that there is no chance for mismatch between the host
apparmor userspace and this snapd internal one.