gadget/many: rm, delay sector size + structure size checks to runtime
We don't have a good way to know ahead of time the structure size of a disk, and
encoding it in the gadget is not a good solution, since it means that you need a
new gadget for a new disk with a different sector size when everything else
about the gadget is the same.

As such, don't assume anything about the disk volume's sector size before we
have available to us the specific sector size of the disk at runtime during UC20
install mode, and at that point verify that the structure sizes in the gadget
are even multiples of the sector size.

This will allow gadgets with structure sizes that are multiples of 4096 to be
used on disks with sector sizes of either 512 or 4096, whereas before we would
fail to install on disks with sector size of 4096.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>