snap: add deprecation noticed to "snap run --gdb"
Now that we promoted "snap run --gdbserver" the "snap run --gdb"
option should be deprecated. We plan to not improve it and it
needs to run as root, the program and gdb share a pty and it's
overall not great. So let's mark it deprecated and point users
to `snap run --gdbserver` instead.