cmd/snap-bootstrap: check for a tried recovery system and reboot accordingly
When we 'try' a new recovery system, make sure it boots up to a point where
snap-bootstrap has set up the system for further execution in after-pivot world.
This ensures that the try system is functional until that point, but does not
execute any services from it in case those would alter the system state.

We also verify that the ubuntu-data is mounted and accessible, and as such the
new recovery system should be functional if fully booted.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>