interfaces/docker-support: add autobind unix rules to docker-support
This will enable new versions of docker to use the new version of go-systemd
which communicates with journald via an automatic/empty address specification on
a SOCK_DGRAM socket.

This access already exists in the kubernetes-support interface as the
autobind-unix flavor, so update that comment to reflect that we should only use
the new syntax when AppArmor userspace is vendored into snapd.

See also https://github.com/coreos/go-systemd/issues/331

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>