usersession/autostart: mv perms fixup to cmd_userd.go
This will make it more clear we are doing this, doing this from inside the
function named AutostartSessionApps was a bit weird, here it's more visible.

This does require us to export mocking the userCurrent function for the
autostart package, since otherwise the cmd_userd_test.go tests fail as they try
to use the real user running the test's home directory and not the mocked one.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>