overlord: ignoring non-kernel assets when updating the kernel
This commit changes the UpdatePolicyFunc to allow modifying the
"from" and "to" structures. This is needed when a kernel update
is performed and only assets from the kernel need updating and
the parts from the gadget should stay untouched on disk. Without
this the gadget content would be rewriten as well.