devicestate: kernel asset update always uses KernelUpdatePolicy
When doUpdateGadgetAssets() happens for a kernel update, always
use the KernelUpdatePolicy even when a remodel is in progress.

We may reconsider in the future but it's clearer if the kernel
only deals with the kernel bits during the refresh and not
rewrites the gadget (which may at this point still the old
gadget).