packaging: drop 99-snapd.conf via dpkg-maintscript-helper
We dropped shipping /etc/sudoers.d/99-snapd.conf in PR#8885 but
did not clean up properly in the debian/* packaging. This is a
conffile in debian packaging terms so it needs special handling.

Fortunately the dpkg-maintscript-helper makes this easy(ish)
now with debian/<pkg>.maintscript. Note that the version in
the maintscript files is 2.50~ because we need to use the
version of the package with the fix, not the version where
the conffile was dropped.

This fixes LP: #1915156