cmd/snap-bootstrap/initramfs-mounts: move time forward using ref files
Move the time of the system forward in the initrd monotonically, using a kernel
initrd timestamp file which will be included in the filesystem of the initrd and
is thus signed/trusted, as well as using the modification time of the systems
directory in ubuntu-seed.

Don't move the time forward at all if we don't have the kernel initrd timestamp
file to protect against doing anything unexpected on old kernels that happen to
get built with this version of snap-bootstrap, and also don't ever move the time
backwards if the current time the kernel already has is newer than the 
timestamps in the initrd file or the ubuntu-seed systems directory.

Partially fixes: https://bugs.launchpad.net/snapd/+bug/1878422

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>