boot/flags.go: add NextBootFlags() for getting snapd_next_boot_flags bootenv
This function will be used by snap system-mode to get the current boot flags
from the bootenv - which may be for this current boot or may be for the next
boot depending on the context in which the function is called.

The only current flag we will support is "factory", but to be robust, this var
can be a comma separated list, but with the caveat that it should be kept to the
maximum length of a lk bootenv variable value which is 256 characters (including
a NUL at the end).

The "factory" boot flag will mainly be set from ubuntu-image, and be cleared
during install mode at the end.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>