tests/main/lxd/prep-snapd-in-lxd.sh: dump contents of sources.list
We _still_ continue to see errors like

+ apt update

WARNING: apt does not have a stable CLI interface. Use with caution in scripts.

E: Type 'xenial-security' is not known on line 50 in source list /etc/apt/sources.list
E: The list of sources could not be read.

executing this test, so rather than just be confused about what apt says is
wrong about the file, instead just dump that file before every apt operation, in
the hopes that it eventually fails again and we can see what the output was
right before the apt operation fails.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>