snapstate: error is "snap refresh" has no updates because of conflicts
When a `snap refresh` is running a subsequent `snap refresh`
will report:
```
All snaps are up to date.
```

This is misleading and caused confusion and bugreports by
customers.

This PR fixes this by reporting that some refreshes could not be
done because of conflicting changes. Something like:
```
error: cannot refresh: auto-refresh has conflicting changes in progress
```

Happy to tweak the message.