tests/lib/prepare.sh: use /etc/group and friends from the core20 snap
Using /etc/group and other things from the running system we are doing the
surgery on instead of the core20 snap leads to an incorrect permissions setup,
where at a minimum the permissions of the dbus helper are incorrect and trying
to D-Bus activate a service fails with

The permission of the setuid helper is not correct

on /usr/lib/dbus-1.0/dbus-daemon-launch-helper. Specifically, the gid 103 on
UC20 is supposed to be messagebus, but on classic 20.04, gid 103 is
systemd-network, leading the dbus daemon to correctly think it has the wrong
permissions.

This allows us to re-enable some UC20 tests which were initially disabled and
never re-enabled.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>