boot/assets.go: add InstallObserverOptions to allow providing previous assets
While working on splitting up the current set of operations from the
setup-run-system task into multiple tasks, we need to serialize the
TrustedAssetInstallObserver to state.json. Serializing a
TrustedAssetInstallObserver requires storing out-of-band the set of tracked
assets for the run and recover mode bootloader, so export the functions which
return the current set of tracked assets, and also allow providing the
previously serialized AssetsMap as input to TrustedAssetsInstallObserverForModel
via InstallObserverOptions.

We can also more naturally express whether or not to use encryption via a
setting in the Options struct too, so move that parameter to
InstallObserverOptions too.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>