o/devicestate/devicemgr: register install-device hook, run if present in install
Run a new install-device hook during install mode after setup-run-system, but
before the ensure-next-boot-to-run-mode task.

Also fix a subtle bug in the addTask() function, it wasn't setting the previous
task properly, so all new tasks would just depend on the original
setup-run-system task, and not on the next task in the chain.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>