packaging: drop dh-systemd from build-depends on ubuntu-16.04+
The dh-systemd functionality got merged into debhelper with
version 9.20160709. So this commit expresses this by changing
the build-depends on `debhelper (>= 9.20160709~) | dh-systemd`

This will ensure that on Ubuntu 16.04 where the debhelper
version is lower (9.20160115ubuntu3) we still get dh-systemd
but for all newer versions of Ubuntu we only install debhelper
instead.

Note that we already dropped dh-systemd in our Debian packaging
branch.