cmd/snap-bootstrap/initramfs-mounts: move time forward using assertion times
This uses ReadSystemEssentialAndBetterEarliestTime in order to get a trusted
time that the system can be moved forward to, providing as input the current
time from the system, which should be trusted, as it is either from a RTC, or
from systemd moving the time forward to when the initrd was built.

Also add tests for the various different relative times that can exist during
the initramfs for each mode.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>