wrappers: add dependency on usr-lib-snapd.mount for services on core with snapd snap
In a core system with snapd snap, the contents of <snapd-snap>/usr/lib/snapd are
exported to the host via a bind mount of done by usr-lib-snapd.mount unit. The
mount unit implicitly depends on the snap mount to be present. By adding that
tooling mount as a dependency for all services on core where snapd is present we
gain some confidence that by the time the unit is started:
- /usr/lib/snapd contains the right tooling
- /snap/snapd/<rev> is mounted

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>