wrappers: use /usr/lib/snapd/snap when on core with snapd snap
On Core systems with the snapd snap, /usr/bin/snap is a symlink pointing to
/snap/snapd/current/usr/bin/snap. However, while the snapd snap is being
refreshed, the current symlink becomes invalid for a short while. Should a
reboot occur while this happens, the existing snap services will be unable to
start as /usr/bin/snap points to an invalid location.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>