overlord/snapstate/backend: do not remove current symlink for the snapd snap
When snap is refreshed, the current symlink would briefly go away between
unlink-current-snap and link-snap tasks. If the system gets rebooted at this
time, during startup all services from snaps will fail to start. This is caused
by the fact that /usr/bin/snap, which on Core is linked to
/snap/snapd/current/usr/bin/snap will be dangling, as /snap/snapd/current was
removed.

Address the problem by not removing current of the snapd snap, unless unlinking
is called during removal for the first install of the snapd snap on core.
However, even in such scenario, the /usr/bin/snap will be invoked from the core
snap.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>