o/devicestate: fix install mode change tests and rename task
Fix the NoRestarts variant of the test that ensures that the setup-run-system
task no longer restarts the device, and instead leaves that to the (newly
renamed) restart-system-to-run-mode task.

Rename ensure-next-boot-to-run-mode to restart-system-to-run-mode as this better
encapsulates what the change does, in that it reboots the system.

Also remove a redundant state setting of setup-run-system-task that was
necessary in a previous implementation but is no longer used for anything.

Finally, fix a subtle bug in the addTask function which was not moving the prev
task forward with each new task. this didn't manifest itself here, but was
manifesting itself in follow-up branches so fix it here before we forget.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>