overlord/snapstate, wrappers: add dependency on usr-lib-snapd.mount for services on core with snapd snap
* snapdtool: add helper for checking whether the current process runs from the snapd snap

Add a helper that checks whether the currently process binary is running form
the snapd snap.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>

* wrappers: add dependency on usr-lib-snapd.mount for services on core with snapd snap

In a core system with snapd snap, the contents of <snapd-snap>/usr/lib/snapd are
exported to the host via a bind mount of done by usr-lib-snapd.mount unit. The
mount unit implicitly depends on the snap mount to be present. By adding that
tooling mount as a dependency for all services on core where snapd is present we
gain some confidence that by the time the unit is started:
- /usr/lib/snapd contains the right tooling
- /snap/snapd/<rev> is mounted

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>

* Revert "snapdtool: add helper for checking whether the current process runs from the snapd snap"

This reverts commit 93fa5a3f0c91860cd5e0a54f0d364c7a3d73e0a0.

* wrappers: add flag for expressing a dependency on the snapd tooling

Add a flag to the add services helper that indicates the services require
additional dependency on the snapd tooling provided by the snapd snap.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>

* overlord/snapstate: pass flags to require snapd tooling when linking a snap on UC with base

When linking a snap on an Ubuntu Core system which uses one of core* bases, set
the additional flags that require dependency on the snapd tooling mount.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>

* overlord/snapstate, wrappers: tweak naming

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>

* overlord/snapstate: comments and test tweaks

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>