o/configstate/configcore/picfg.go: use ubuntu-seed config.txt in uc20 run mode
In UC20 run mode, when applying pi-config changes, we want to apply them to the
config.txt that exists on the ubuntu-seed partition instead of the ubuntu-boot
partition since the current pi bootloader setup is such that ubuntu-boot does
not contain any boot assets except the kernel and initrd, and all gpu
bootloader assets like config.txt actually live on ubuntu-seed.

This change does not fix applying gadget defaults to the config.txt before the
device is booted, that will be a separate change that will build on this.

Also, move the mocking of an empty command line up one level to the 
configcoreSuite so that all config operations have an empty kernel command line
mocked.

Fixes: https://bugs.launchpad.net/snapd/+bug/1909353

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>