o/configstate/configcore/picfg.go: refactor piConfigFile, use UC20Recovery opt
Use the UC20Recovery option to decide what sub directory to look for config.txt
in. Specifically, when UC20Recovery is true, then we know the directory we are
working on is a recovery system, and so the config.txt file is in the root of
that recovery system.

Additionally, fix up the handling of the location of the config.txt for UC20
run mode in the case where there is a RootDir provided in the fsOnlyContext. In
UC20 run mode when we are using the fsOnlyContext, the RootDir provided there
is actually still wrong and should be ignored in favor of /run/mnt/ubuntu-seed
due to decisions made that there is only one config.txt for a UC20 pi system,
and it lives on the ubuntu-seed partition, and thus all config modifies that
version. Also add a unit test for this case.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>