packaging: stop snapd.{socket,service} before purging data
This commit will ensure that snapd is properly stopped before
the data is removed. If we don't do that there is a race when
e.g. a refresh is in progress while "dpkg --purge snapd" is
running. The postrm will remove a bunch of data but because
snapd is still running it will write out some files like
state.json.