o/configstate/configcore: explicitly ignore install, recover mode pi-config
It is currently unclear what the desired or intended behavior should be when
applying pi-config in install or recover modes, since the underlying config.txt
could have been changed at runtime by `snap set pi-config.xyz=foo`, in which
case applying gadget defaults for pi-config when seeding the ephemeral system
in UC20 recover mode would be undoing that if it wrote to the actual
config.txt on ubuntu-seed. Additionally, even if it did write to it, no changes
to config.txt take place until the reboot, at which point we are no longer in
recover mode (or install mode for that matter) and so applying the config in
these modes currently doesn't provide any value, so just ignore these modes
entirely.

We may someday revisit this with something like per-mode gadget config, or
apply config for a recovery system somehow before booting into it if we are
able, etc. but for now we don't.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>