boot/flags: make BootFlags filter the result + return specific error
When the flags are filtered out to only be the ones that are currently
understood, then we return a specific error type that can be checked for in
addition to the full set of understood boot flags. This allows more robust
implementations to process understood/supported boot flags while still logging
unknown ones that pop up from untrusted places like the bootenv or possibly a
newer snapd that we reverted away from (or one that was newer and running in
run mode, but now we are running in recover mode).

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>