interfaces/builtin: introduce dsp interface
This interface will be used by specific devices with plugs that have flavor
attributes identifying which will provide board specific access to digital
signal processor devices which are very specific to a particular BSP or board.

The interface is meant to be provided by gadget snap slots, with the particular
flavor identified by the gadget slot attribute.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>