tests/lib/prepare-restore.sh: wait for the abort to be done with `snap watch`
We have seen now instances where the auto-refresh change is aborted, but it
didn't finish by the time we get below and try to remove snaps, so watch for
the change to be done before we proceed to trying to remove the snaps. See for
example this output:

++ grep Do
++ grep -v Done
++ tail -n +2
++ awk '{print $1}'
++ snap changes
+ for chg in $(snap changes | tail -n +2 | grep Do | grep -v Done | awk '{print $1}')
+ snap abort 3
++ awk '{print $1,$6}'
++ grep -Po '(.+)\s+.*(?!base).*'
++ tail -n +2
++ snap list
++ awk '{print $1}'
+ for sn in $(snap list | tail -n +2 | awk '{print $1,$6}' | grep -Po '(.+)\s+.*(?!base).*' | awk '{print $1}')
+ '[' core18 '!=' snapd ']'
+ '[' core18 '!=' core ']'
+ snap remove core18
error: snap "core18" has "auto-refresh" change in progress
+ true
+ for sn in $(snap list | tail -n +2 | awk '{print $1,$6}' | grep -Po '(.+)\s+.*(?!base).*' | awk '{print $1}')
+ '[' google-cloud-sdk '!=' snapd ']'
+ '[' google-cloud-sdk '!=' core ']'
+ snap remove google-cloud-sdk
error: snap "google-cloud-sdk" has "auto-refresh" change in progress
+ true
+ for sn in $(snap list | tail -n +2 | awk '{print $1,$6}' | grep -Po '(.+)\s+.*(?!base).*' | awk '{print $1}')
+ '[' lxd '!=' snapd ']'
+ '[' lxd '!=' core ']'
+ snap remove lxd
error: snap "lxd" has "auto-refresh" change in progress
+ true
+ for sn in $(snap list | tail -n +2 | awk '{print $1,$6}' | grep -Po '(.+)\s+.*(?!base).*' | awk '{print $1}')
+ '[' snapd '!=' snapd ']'
+ distro_purge_package snapd

The lxd, google-cloud-sdk, and core18 snaps were all being refreshed by change
3, but it wasn't done being aborted before we tried to remove the snaps.

Also fix the regexp so it doesn't try to remove the core18 snap, which is a
base and can't be removed until after all the app snaps are removed.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>